MediaCodec - use software decoder/encoder h264
-Either our OMX patch is incompleted or our h.264 encoder/decoders
do not properly implement functions of the MediaCodec class
introduced in API 16 (JellyBean). So the solution here
is to use software encoder/decoder for video/avc.

NOTE: this also requires a patch in libstagefright to skip codec
configuration when a software encoder/decoder is detected and
to add software encoders to the media_codecs.xml

Change-Id: I787057fa39e6b674e832a3f762436e1a5403f6d2